fix bugs; fully implement loot

+++ src/game/objects/loot.ts
@@ -1,12 +1,4 @@
-import {
-    CollisionCategory,
-    Constants,
-    IdToGameType,
-    ObjectKind,
-    removeFrom,
-    type SurvivBitStream,
-    TypeToId
-} from "../../utils";
+import { CollisionCategory, Constants, ObjectKind, removeFrom, type SurvivBitStream, TypeToId } from "../../utils";
 import { type Game } from "../game";
 import { GameObject } from "../gameObject";
 import { type Player } from "./player";
@@ -28,12 +20,18 @@ export class Loot extends GameObject {
         super(id, typeString, position, layer, undefined, game);
         this.kind = ObjectKind.Loot;
         this.count = count;
-        this.body = game.world.createBody({ position });
+        this.body = game.world.createBody({
+            type: "dynamic",
+            position,
+            fixedRotation: true
+        });
         this.body.createFixture({
             shape: Circle(position, 1),
+            restitution: 0.0,
+            density: 10.0,
+            friction: 0.0,
             filterCategoryBits: CollisionCategory.Loot,
-            filterMaskBits: CollisionCategory.Obstacle | CollisionCategory.Player,
-            filterGroupIndex: 3
+            filterMaskBits: CollisionCategory.Obstacle
         });
     }
 
@@ -41,65 +39,40 @@ export class Loot extends GameObject {
         return this.body!.getPosition();
     }
 
-    interact(p: Player): void { // TODO Clean up, remove redundant code
+    interact(p: Player): void {
         let result: PickupMsgType = PickupMsgType.Success;
+        let deleteItem = true;
+        let playerDirty = false;
         if(this.typeString.endsWith("scope")) {
             if(p.inventory[this.typeString] > 0) result = PickupMsgType.AlreadyEquipped;
             else {
+                p.inventory[this.typeString]++;
                 p.activeItems.scope.typeString = this.typeString;
                 p.activeItems.scope.typeId = TypeToId[this.typeString];
                 if(p.isMobile) p.zoom = Constants.scopeZoomRadius.mobile[this.typeString];
                 else p.zoom = Constants.scopeZoomRadius.desktop[this.typeString];
-                p.zoomDirty = true;
             }
         } else if(this.typeString.startsWith("backpack")) {
-            const packLevel = parseInt(this.typeString.charAt(9)); // Last digit of the ID is always the item level
-            if(packLevel < p.packLevel) result = PickupMsgType.BetterItemEquipped;
-            else if(packLevel === p.packLevel) result = PickupMsgType.AlreadyEquipped;
-            else {
-                const noPack: boolean = (p.packLevel === 0);
-                p.packLevel = packLevel;
-                if(!noPack) {
-                    const oldItem: Loot = new Loot(this.game!.nextObjectId, IdToGameType[450 + p.packLevel], this.position, this.layer, this.game!, 1);
-                    this.game!.objects.push(oldItem);
-                    this.game!.fullDirtyObjects.push(oldItem);
-                }
-            }
+            result = this.pickUpTieredItem("backpack", p);
+            playerDirty = true;
         } else if(this.typeString.startsWith("chest")) {
-            const chestLevel = parseInt(this.typeString.charAt(6));
-            if(chestLevel < p.chestLevel) result = PickupMsgType.BetterItemEquipped;
-            else if(chestLevel === p.chestLevel) result = PickupMsgType.AlreadyEquipped;
-            else {
-                const noChest: boolean = (p.chestLevel === 0);
-                p.chestLevel = chestLevel;
-                if(!noChest) {
-                    const oldItem: Loot = new Loot(this.game!.nextObjectId, IdToGameType[457 + p.chestLevel], this.position, this.layer, this.game!, 1);
-                    this.game!.objects.push(oldItem);
-                    this.game!.fullDirtyObjects.push(oldItem);
-                }
-            }
+            result = this.pickUpTieredItem("chest", p);
+            playerDirty = true;
         } else if(this.typeString.startsWith("helmet")) {
-            const helmetLevel = parseInt(this.typeString.charAt(7));
-            if(helmetLevel < p.helmetLevel) result = PickupMsgType.BetterItemEquipped;
-            else if(helmetLevel === p.helmetLevel) result = PickupMsgType.AlreadyEquipped;
-            else {
-                const noHelmet: boolean = (p.helmetLevel === 0);
-                p.helmetLevel = helmetLevel;
-                if(!noHelmet) {
-                    const oldItem: Loot = new Loot(this.game!.nextObjectId, IdToGameType[453 + p.helmetLevel], this.position, this.layer, this.game!, 1);
-                    this.game!.objects.push(oldItem);
-                    this.game!.fullDirtyObjects.push(oldItem);
-                }
-            }
+            result = this.pickUpTieredItem("helmet", p);
+            playerDirty = true;
         } else {
             const currentCount: number = p.inventory[this.typeString];
-            const maxCapacity: number = Constants.bagSizes[this.typeString][p.packLevel];
-            if(currentCount + 1 > maxCapacity) result = PickupMsgType.Full;
-            else if(currentCount + this.count > maxCapacity) {
-                const increase = (currentCount + this.count) - maxCapacity;
-                (p.inventory[this.typeString] as number) += increase;
-                this.count -= increase;
-                result = PickupMsgType.Success;
+            const maxCapacity: number = Constants.bagSizes[this.typeString][p.backpackLevel];
+            if(currentCount + this.count <= maxCapacity) {
+                (p.inventory[this.typeString] as number) += this.count;
+            } else if(currentCount + 1 > maxCapacity) {
+                result = PickupMsgType.Full;
+            } else if(currentCount + this.count > maxCapacity) {
+                (p.inventory[this.typeString] as number) = maxCapacity;
+                this.count = (currentCount + this.count) - maxCapacity;
+                this.game!.fullDirtyObjects.push(this);
+                deleteItem = false;
             }
         }
 
@@ -107,15 +80,37 @@ export class Loot extends GameObject {
             p.sendPacket(new PickupPacket(this.typeString!, this.count, result!));
         }
         if(result! === PickupMsgType.Success) {
-            removeFrom(this.game!.objects, this);
-            this.game!.deletedObjects.push(this);
-            this.game!.world.destroyBody(this.body!);
-            this.interactable = false;
-
-            this.game?.fullDirtyObjects.push(p);
-            p.fullDirtyObjects.push(p);
+            if(deleteItem) {
+                removeFrom(this.game!.objects, this);
+                //removeFrom(this.game!.loot, this);
+                this.game!.deletedObjects.push(this);
+                this.game!.world.destroyBody(this.body!);
+                this.interactable = false;
+            }
+            if(playerDirty) {
+                this.game?.fullDirtyObjects.push(p);
+                p.fullDirtyObjects.push(p);
+            }
             p.inventoryDirty = true;
         }
+    }
+
+    private pickUpTieredItem(type: string, p: Player): PickupMsgType {
+        const oldLevel: number = p[`${type}Level`];
+        const newLevel: number = parseInt(this.typeString.charAt(this.typeString.length - 1)); // Last digit of the ID is the item level
+        if(newLevel < oldLevel) return PickupMsgType.BetterItemEquipped;
+        else if(newLevel === oldLevel) return PickupMsgType.AlreadyEquipped;
+        else {
+            p[`${type}Level`] = newLevel;
+            if(oldLevel !== 0) { // If oldLevel === 0, the player didn't have an item of this type equipped, so don't drop loot
+                // Example: if type = helmet and p.helmetLevel = 1, typeString = helmet01
+                const oldItem: Loot = new Loot(this.game!.nextObjectId, `${type}0${oldLevel}`, this.position, this.layer, this.game!, 1);
+                this.game!.objects.push(oldItem);
+                //this.game!.loot.push(oldItem);
+                this.game!.fullDirtyObjects.push(oldItem);
+            }
+        }
+        return PickupMsgType.Success;
     }
 
     serializePartial(stream: SurvivBitStream): void {
